refactor(share-target): fold upload action into CreatePurchaseInvoice and move bridge JS
CreatePurchaseInvoice now implements HandlesSharedFiles directly instead of
a separate UploadPurchaseInvoice wrapper, so there is one action per concept.

The inline Capacitor logic in the share-target view moves into nuxbe-bridge.js
as loadSharedFiles/clearSharedFiles, exposed via window.nuxbeShareTarget. The
helpers resolve the Capacitor plugins on each call, so the page can load before
the native bridge is ready. The Blade view is now a thin state machine.

- Remove UploadPurchaseInvoice and its test
- Drop redundant doc comments from the contract and trait
